% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liuReg.R
\name{plot.liuReg}
\alias{plot.liuReg}
\title{Diagnostic plots for a liuReg object}
\usage{
\method{plot}{liuReg}(x, which = 1:2, ...)
}
\arguments{
\item{x}{a \code{liuReg} object.}

\item{which}{subset of 1:2.}

\item{...}{unused.}
}
\description{
Diagnostic plots for a liuReg object
}
