% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raiseReg.R
\name{plot.raiseReg}
\alias{plot.raiseReg}
\title{Diagnostic plots for a raiseReg object}
\usage{
\method{plot}{raiseReg}(x, which = 1:4, ...)
}
\arguments{
\item{x}{a \code{raiseReg} object.}

\item{which}{subset of 1:4 selecting which plots to draw.}

\item{...}{passed on to the underlying plotting functions.}
}
\description{
Produces the same family of goodness-of-fit / diagnostic plots as
\code{plot.lm}: residuals vs fitted, a normal Q-Q plot of the residuals,
a scale-location plot, and a before/after bar chart of VIF values.
}
