% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robLiu.R
\name{plot.robLiu}
\alias{plot.robLiu}
\title{Diagnostic plots for a robLiu object}
\usage{
\method{plot}{robLiu}(x, which = 1:2, ...)
}
\arguments{
\item{x}{a \code{robLiu} object.}

\item{which}{subset of 1:2.}

\item{...}{unused.}
}
\value{
Invisibly returns the fitted \code{robLiu} object `x`.
  Called for its side effect of drawing diagnostic plots.
}
\description{
Diagnostic plots for a robLiu object
}
