% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robRidge.R
\name{plot.robRidge}
\alias{plot.robRidge}
\title{Diagnostic plots for an robRidge object}
\usage{
\method{plot}{robRidge}(x, which = 1:2, ...)
}
\arguments{
\item{x}{an \code{robRidge} object.}

\item{which}{subset of 1:2.}

\item{...}{unused.}
}
\description{
Diagnostic plots for an robRidge object
}
