% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeReg.R
\name{plot.ridgeReg}
\alias{plot.ridgeReg}
\title{Diagnostic plots for a ridgeReg object}
\usage{
\method{plot}{ridgeReg}(x, which = 1:2, ...)
}
\arguments{
\item{x}{a \code{ridgeReg} object.}

\item{which}{subset of 1:2.}

\item{...}{unused.}
}
\description{
Diagnostic plots for a ridgeReg object
}
